// vscode-dotnet-runtime-library/src/EventStream/EventType.d.ts
export declare enum EventType {
    DotnetAcquisitionStart = 0,
    DotnetSDKAcquisitionStart = 1,
    DotnetRuntimeAcquisitionStart = 2,
    DotnetAcquisitionCompleted = 3,
    DotnetAcquisitionError = 4,
    DotnetAcquisitionSuccessEvent = 5,
    DotnetAcquisitionMessage = 6,
    DotnetAcquisitionTest = 7
}